Rename mutations files for consistency
