+++ kolibri/plugins/management/assets/src/device_management/state/mutations/manageContentMutations.js
